ebuild: refactor flushing vdb keys
Refactor how we flush these VDB keys in build-info/ to make it easier
to implement bug #913628: it was a pain with the forced-append of a newline
even if we might tamper with the contents later on, so just postpone flushing
to disk until the end. It saves some repetition too, so double win.

Bug: https://bugs.gentoo.org/913628
Signed-off-by: Sam James <sam@gentoo.org>